chore(wizard-refactor): extract tools/create_institution
Pinned by 2 dispatcher tests (short-circuit + name-match). 28/28 green.
